feature: new option cache-mem-size

+++ src/include/atomic.h
@@ -29,7 +29,7 @@
  * Atomic type.
  */
 typedef struct {
-	int counter;
+	long counter;
 } atomic_t;
 
 #define ATOMIC_INIT(i)  { (i) }
@@ -40,7 +40,7 @@ typedef struct {
  *
  * Atomically reads the value of @v.
  */
-static inline int atomic_read(const atomic_t *v)
+static inline long atomic_read(const atomic_t *v)
 {
 	return READ_ONCE((v)->counter);
 }
@@ -50,7 +50,7 @@ static inline int atomic_read(const atomic_t *v)
  * @param v pointer of type atomic_t
  * @param i required value
  */
-static inline void atomic_set(atomic_t *v, int i)
+static inline void atomic_set(atomic_t *v, long i)
 {
     v->counter = i;
 }
@@ -60,7 +60,7 @@ static inline void atomic_set(atomic_t *v, int i)
  * @param i integer value to add
  * @param v pointer of type atomic_t
  */
-static inline void atomic_add( int i, atomic_t *v )
+static inline void atomic_add( long i, atomic_t *v )
 {
 	(void)__sync_add_and_fetch(&v->counter, i);
 }
@@ -72,7 +72,7 @@ static inline void atomic_add( int i, atomic_t *v )
  *
  * Atomically subtracts @i from @v.
  */
-static inline void atomic_sub( int i, atomic_t *v )
+static inline void atomic_sub( long i, atomic_t *v )
 {
 	(void)__sync_sub_and_fetch(&v->counter, i);
 }
@@ -86,7 +86,7 @@ static inline void atomic_sub( int i, atomic_t *v )
  * true if the result is zero, or false for all
  * other cases.
  */
-static inline int atomic_sub_and_test( int i, atomic_t *v )
+static inline long atomic_sub_and_test( long i, atomic_t *v )
 {
 	return !(__sync_sub_and_fetch(&v->counter, i));
 }
@@ -120,7 +120,7 @@ static inline void atomic_dec( atomic_t *v )
  *
  * Atomically increments @v by 1.
  */
-static inline int atomic_inc_return( atomic_t *v )
+static inline long atomic_inc_return( atomic_t *v )
 {
 	return __sync_add_and_fetch(&v->counter, 1);
 }
@@ -132,7 +132,7 @@ static inline int atomic_inc_return( atomic_t *v )
  * Atomically decrements @v by 1.  Note that the guaranteed
  * useful range of an atomic_t is only 24 bits.
  */
-static inline int atomic_dec_return( atomic_t *v )
+static inline long atomic_dec_return( atomic_t *v )
 {
 	return __sync_sub_and_fetch(&v->counter, 1);
 }
@@ -145,7 +145,7 @@ static inline int atomic_dec_return( atomic_t *v )
  * returns true if the result is 0, or false for all other
  * cases.
  */
-static inline int atomic_dec_and_test( atomic_t *v )
+static inline long atomic_dec_and_test( atomic_t *v )
 {
 	return !(__sync_sub_and_fetch(&v->counter, 1));
 }
@@ -158,7 +158,7 @@ static inline int atomic_dec_and_test( atomic_t *v )
  * and returns true if the result is zero, or false for all
  * other cases.
  */
-static inline int atomic_inc_and_test( atomic_t *v )
+static inline long atomic_inc_and_test( atomic_t *v )
 {
 	return !(__sync_add_and_fetch(&v->counter, 1));
 }
@@ -172,7 +172,7 @@ static inline int atomic_inc_and_test( atomic_t *v )
  * if the result is negative, or false when
  * result is greater than or equal to zero.
  */
-static inline int atomic_add_negative( int i, atomic_t *v )
+static inline long atomic_add_negative( long i, atomic_t *v )
 {
 	return (__sync_add_and_fetch(&v->counter, i) < 0);
 }
